# automatic cross platform progress monitoring