@pfitzseb @tkf  @devmotion @c42f any issues you see with doing this kind of thing?