@ChrisRackauckas 2020-03-09
No noticable overhead:

```julia
using OrdinaryDiffEq
using BenchmarkTools

@btime solve(
    ODEProblem((u, p, t) -> -u, 1.0, nothing),
    Euler();
    dt = 0.5,
    tspan = (0.0, 0.1),
    #progress = true,
    #progress_steps = 1,
)

# Before
44.600 μs (334 allocations: 26.61 KiB)
45.900 μs (334 allocations: 26.61 KiB)

# After
45.000 μs (338 allocations: 26.75 KiB)
46.500 μs (338 allocations: 26.75 KiB)
```