@ChrisRackauckas 2020-03-09
@davidanthoff is there a logger that needs to be set with VSCode?