@davidanthoff 2020-03-09
We don't have any progress UI support in VS Code.