@c42f 2020-03-09
I don't think this is a good solution.

Setting logger sinks should be controlled at the top application level and never in libraries. This allows libraries to compose together while giving the application control of the log stream. (What if people want to send informational messages from DiffEq to a file? (Or from a user defined function DiffEq calls?) Overriding the global logger sink prevents this.

It also forces all users of `DiffEqBase` to install Atom and Juno which seems far from ideal.

In the case of interactive work, the "application level" for the purposes of user code is the REPL in use. So it's appropriate for
* Juno to provide a default logger (and a way to configure a custom one)
* The standard REPL to provide a default logger, and allow users to override it in their `startup.jl`
* VSCode to provide a default logger (and a way to configure it)

In summary, if you're having trouble with Juno + TerminalLoggers, we need to figure out what's going wrong and sort it out at the application level. Not in the libraries which the user may happen to have loaded.