@ChrisRackauckas 2020-03-09
>It also forces all users of DiffEqBase to install Atom and Juno which seems far from ideal.

No, it definitely does not do that...

>Setting logger sinks should be controlled at the top application level and never in libraries. This allows libraries to compose together while giving the application control of the log stream. What if people want to send informational messages from DiffEq to a file? (Or from a user defined function DiffEq calls?) Overriding the global logger sink prevents this.

What about having the logger as a keyword argument? I'm looking for a solution for today, and a discussion for tomorrow.