@tkf 2020-03-09
I agree with all what @c42f said. But, for a _short_ term solution, I don't think it's a horrible idea.

> What if people want to send informational messages from DiffEq to a file?

One workaround is to use `LoggingExtras.TeeLogger` like @devmotion is doing:

https://github.com/TuringLang/AbstractMCMC.jl/blob/4bfe617eca199852251f6314627d7eba242e6294/src/AbstractMCMC.jl#L38-L48