@c42f 2020-03-09
> > It also forces all users of DiffEqBase to install Atom and Juno which seems far from ideal.
> 
> No, it definitely does not do that...

Ok. But I'm then confused that `Juno` seems to be included in the Project.toml here. Is it just the Juno backend which is relatively lightweight?