@ChrisRackauckas 2020-03-09
>Ok. But I'm then confused that Juno seems to be included in the Project.toml here. Is it just the Juno backend which is relatively lightweight?

Yeah, it's just some functions for interacting with Juno, but doesn't include Atom.jl or the big dependencies.