@tkf 2020-03-09
Actually, I think a better solution is to just check `progress` keyword argument and print some helpful message:

```julia
if get(NamedTuple(kwargs), :progress, false) === true &&
   Logging.min_enabled_level(current_logger()) > LogLevel(-1)
    @warn """
    `progress = true` is passed but your logger does not support progress bars.
    Please see the documentation for how to set it up:
    https://...
    """
end
```