@tkf 2020-03-09
Maybe we can add a function (say) `ProgressLogging.enabled_for(logger)` so that we don't need a hack like `Logging.min_enabled_level(current_logger()) > LogLevel(-1)`.