@c42f 2020-03-09
> Actually, I think a better solution is to just check progress keyword argument and print some helpful message:

That would also work (add the `maxlog=1` keyword to that `@warn`, I guess).

I think the long term direction for progress should be to remove the `progress` keyword from APIs like `solve` and have a neat generic way to configure progress filters in the logging backends.

> Maybe we can add a function (say) `ProgressLogging.enabled_for(logger)` so that we don't need a hack like `Logging.min_enabled_level(current_logger()) > LogLevel(-1)`.

This seems like a good thing to try out. I feel like it's roughly the job of `shouldlog(logger, level, ...)`, but we don't have enough sophistication in the semantics of `shouldlog` to express this right now. So having a `ProgressLogging.enabled_for` could do the trick for now.

Btw, I was thinking something kind of related — that a filter for progress logs could sensibly live in `ProgressLogging` itself, keeping all the `Progress` related stuff more self-contained. This is somewhat the thing that `ProgressLogging.asprogress` does right now.