@tkf 2020-03-09
> add the `maxlog=1` keyword to that `@warn`

Good idea :+1:

> remove the `progress` keyword from APIs like `solve`

Agree in general, but for DiffEq API, I think it's reasonable to have `progress` argument.  It is a common situation where you want every bit of instructions to be used for solving the problem and not wasting time for creating progress events.

> Btw, I was thinking something kind of related — that a filter for progress logs could sensibly live in `ProgressLogging` itself, keeping all the `Progress` related stuff more self-contained. This is somewhat the thing that `ProgressLogging.asprogress` does right now.

Agree. Let's track that it in https://github.com/JunoLab/ProgressLogging.jl/issues/9