@ChrisRackauckas 2020-03-09
>remove the progress keyword from APIs like solve

I would like this too: everything across Julia should have a common way to turn on/off all logging, like warnings, progress bars, etc. Having a different setup for every package, and not being able to turn it off in many, is just confusing and doesn't compose well.

But for now this seems like the best solution, so I'll merge but I'm glad there's an ongoing discussion to be had 😄