@devmotion 2020-03-09
> I don't think this is a good solution.
>
> Setting logger sinks should be controlled at the top application level and never in libraries. This allows libraries to compose together while giving the application control of the log stream. What if people want to send informational messages from DiffEq to a file? (Or from a user defined function DiffEq calls?) Overriding the global logger sink prevents this.

I completely agree, and so for quite some time I was arguing that we should not provide a default progress logger in AbstractMCMC but only display a warning if the current logger seems unable to handle progress logs (based on the `min_enabled_level` hack). However, I still think it is nice if displaying progress logs just works even without setting any specific progress logger (the main argument being that new users might expect such a functionality from a PPL) if it does not interfere with any custom loggers a user might have defined. Therefore the current approach in AbstractMCMC uses a `LoggingExtras.TeeLogger` and two `LoggingExtras.EarlyFilteredLoggers` to only forward progress logs (or more precisely, logs with log level -1) generated by the current module to the progress logger and forward all other logs to the current logger. The idea was that this allows us to get the best of both worlds - keep the composability of the logging system and provide a default progress logger if it was not set.